JPF:  * Implemented a 'make.section.dirs' on sections.conf so that the SectionManager will auto-create the dirs if needed.  * Fixed a few minor bugs suggested by FindBugs

+++ src/master/src/org/drftpd/jobmanager/JobManager.java
@@ -264,7 +264,7 @@ public class JobManager {
 		}
 		_useCRC = p.getProperty("useCRC", "true").equals("true");
 		_useSSL = p.getProperty("useSSLTransfers", "true").equals("true"); 
-		_sleepSeconds = 1000 * Integer.parseInt(PropertyHelper.getProperty(p,
+		_sleepSeconds = 1000 * Long.parseLong(PropertyHelper.getProperty(p,
 				"sleepSeconds"));
 		if (_runJob != null) {
 			_runJob.cancel();
